refactor: introduce a `PositionedPage` to store a page in the `Scanner`
Update the scanning logic to handle interior pages. Use a depth-first
algorithm to recursively explore the tree, starting from the root page.

Update the `Cursor` to own the payload instead of borrowing it through
the `Pager`. This avoids borrowing the `Pager` mutably from the `Cursor`
and the `Scanner` at the same time, which could lead to lifetime issues
in the future.

The scanning logic maintains a stack of `PositionedPage`s to keep track
of the parent pages which have been visited. At every step there is a
few cases to consider:

* If the current page is a leaf page and the cells haven't been visited
  yet, we'll just have to build a `Cursor` with the current cell's
  payload and return it.
* The current page is an interior page, push the next page (either from
  the current cell or the rightmost pointer) to the stack and continue.
* All cells have been visited of the current page, pop the stack and
  continue from the parent page.